+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Implemented qwerty keyboard (swipe space bar to switch)
 - Implemented horizontal swiping of a key
 - Fixed backspace not working in Termux
+- Increased backspace repeat speed for ASCII
 - Prevented keyboard from showing on main activity startup
 - Added source code link to main activity
 - Improved main activity information text